requirements-full.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
certifi==2023.11.17
requests==2.31.0
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
reportlab==4.0.7
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
plotly==5.16.1